requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.4.2
aiosmtplib==2.0.2
mysql-connector-python==8.2.0
python-multipart==0.0.6